There is one Producer in the producer-consumer problem, Producer is producing some items, whereas there is one Consumer that is consuming the items produced by the Producer. The same memory buffer is shared by both producers and consumers which is of fixed-size. Devise a Java program to producer-consumer problem of operating system using interthread communication function call
There is one Producer in the producer-consumer problem,
Producer is producing some items, whereas there is one
Consumer that is consuming the items produced by the
Producer. The same memory buffer is shared by both
producers and consumers which is of fixed-size. Devise a Java
program to producer-consumer problem of operating system
using interthread communication function call